integer warning fixed

+++ CDOJ-IOS/Function/Time.m
@@ -25,16 +25,16 @@ NSString* getTimeLengthString(NSString* timeStamp) {
     hours %= 24; minutes %= 60; seconds %= 60;
     NSString* str = @"";
     if(days > 0) {
-        str = [NSString stringWithFormat:@"%@%ld %@ ", str, days, days == 1 ? @"day" : @"days"];
+        str = [NSString stringWithFormat:@"%@%ld %@ ", str, (long)days, days == 1 ? @"day" : @"days"];
     }
     if(hours > 0) {
-        str = [NSString stringWithFormat:@"%@%ld %@ ", str, hours, hours == 1 ? @"hour" : @"hours"];
+        str = [NSString stringWithFormat:@"%@%ld %@ ", str, (long)hours, hours == 1 ? @"hour" : @"hours"];
     }
     if(minutes > 0) {
-        str = [NSString stringWithFormat:@"%@%ld %@ ", str, minutes, minutes == 1 ? @"minute" : @"minutes"];
+        str = [NSString stringWithFormat:@"%@%ld %@ ", str, (long)minutes, minutes == 1 ? @"minute" : @"minutes"];
     }
     if(seconds > 0) {
-        str = [NSString stringWithFormat:@"%@%ld %@ ", str, seconds, seconds == 1 ? @"second" : @"seconds"];
+        str = [NSString stringWithFormat:@"%@%ld %@ ", str, (long)seconds, seconds == 1 ? @"second" : @"seconds"];
     }
     return str;
 }
@@ -47,9 +47,9 @@ NSString* getTimeLengthString2(NSString* timeStamp) {
     hours %= 24; minutes %= 60; seconds %= 60;
     NSString* str = @"";
     if(days > 0) {
-        str = [NSString stringWithFormat:@"%@%ld %@ ", str, days, days == 1 ? @"day" : @"days"];
+        str = [NSString stringWithFormat:@"%@%ld %@ ", str, (long)days, days == 1 ? @"day" : @"days"];
     }
-    str = [NSString stringWithFormat:@"%@ %02ld:%02ld:%02ld", str, hours, minutes, seconds];
+    str = [NSString stringWithFormat:@"%@ %02ld:%02ld:%02ld", str, (long)hours, (long)minutes, (long)seconds];
     return str;
 }
